Move API Client to better folder
